+++ gutenberg/packages/core-data/src/selectors.ts
@@ -1,15 +1,8 @@
-/**
- * WordPress dependencies
- */
 import { createSelector, createRegistrySelector } from '@wordpress/data';
 import { addQueryArgs } from '@wordpress/url';
 import type { UndoManager } from '@wordpress/undo-manager';
 import deprecated from '@wordpress/deprecated';
 import type { ConnectionStatus } from '@wordpress/sync';
-
-/**
- * Internal dependencies
- */
 import { STORE_NAME } from './name';
 import {
 	getQueriedItems,
@@ -325,7 +318,7 @@ export interface GetEntityRecord {
 		state: State,
 		kind: string,
 		name: string,
-		key?: EntityRecordKey,
+		recordId?: EntityRecordKey,
 		query?: GetRecordsHttpQuery
 	): EntityRecord | undefined;
 
@@ -336,7 +329,7 @@ export interface GetEntityRecord {
 	>(
 		kind: string,
 		name: string,
-		key?: EntityRecordKey,
+		recordId?: EntityRecordKey,
 		query?: GetRecordsHttpQuery
 	) => EntityRecord | undefined;
 	__unstableNormalizeArgs?: ( args: EntityRecordArgs ) => EntityRecordArgs;
@@ -347,12 +340,12 @@ export interface GetEntityRecord {
  * yet received, undefined if the value entity is known to not exist, or the
  * entity object if it exists and is received.
  *
- * @param state State tree
- * @param kind  Entity kind.
- * @param name  Entity name.
- * @param key   Optional record's key. If requesting a global record (e.g. site settings), the key can be omitted. If requesting a specific item, the key must always be included.
- * @param query Optional query. If requesting specific
- *              fields, fields must always include the ID. For valid query parameters see the [Reference](https://developer.wordpress.org/rest-api/reference/) in the REST API Handbook and select the entity kind. Then see the arguments available "Retrieve a [Entity kind]".
+ * @param state    State tree
+ * @param kind     Entity kind.
+ * @param name     Entity name.
+ * @param recordId Is omitted for keyless entities.
+ * @param query    Optional query. If requesting specific
+ *                 fields, fields must always include the ID. For valid query parameters see the [Reference](https://developer.wordpress.org/rest-api/reference/) in the REST API Handbook and select the entity kind. Then see the arguments available "Retrieve a [Entity kind]".
  *
  * @return Record.
  */
@@ -365,7 +358,7 @@ export const getEntityRecord = createSelector(
 		state: State,
 		kind: string,
 		name: string,
-		key?: EntityRecordKey,
+		recordId?: EntityRecordKey,
 		query?: GetRecordsHttpQuery
 	): EntityRecord | undefined => {
 		logEntityDeprecation( kind, name, 'getEntityRecord' );
@@ -378,14 +371,14 @@ export const getEntityRecord = createSelector(
 
 		if ( ! query || ! query._fields ) {
 			// If expecting a complete item, validate that completeness.
-			if ( ! queriedState.itemIsComplete[ context ]?.[ key ] ) {
+			if ( ! queriedState.itemIsComplete[ context ]?.[ recordId ] ) {
 				return undefined;
 			}
 
-			return queriedState.items[ context ][ key ];
+			return queriedState.items[ context ][ recordId ];
 		}
 
-		const item = queriedState.items[ context ]?.[ key ];
+		const item = queriedState.items[ context ]?.[ recordId ];
 		if ( ! item ) {
 			return item;
 		}
@@ -437,11 +430,11 @@ getEntityRecord.__unstableNormalizeArgs = (
  * Note: This action does not trigger a request for the entity record from the API
  * if it's not available in the local state.
  *
- * @param state State tree
- * @param kind  Entity kind.
- * @param name  Entity name.
- * @param key   Record's key.
- * @param query Optional query.
+ * @param state    State tree
+ * @param kind     Entity kind.
+ * @param name     Entity name.
+ * @param recordId Is omitted for keyless entities.
+ * @param query    Optional query.
  *
  * @return Whether an entity record has been received.
  */
@@ -449,7 +442,7 @@ export function hasEntityRecord(
 	state: State,
 	kind: string,
 	name: string,
-	key?: EntityRecordKey,
+	recordId?: EntityRecordKey,
 	query?: GetRecordsHttpQuery
 ): boolean {
 	const queriedState =
@@ -461,10 +454,10 @@ export function hasEntityRecord(
 
 	// If expecting a complete item, validate that completeness.
 	if ( ! query || ! query._fields ) {
-		return !! queriedState.itemIsComplete[ context ]?.[ key ];
+		return !! queriedState.itemIsComplete[ context ]?.[ recordId ];
 	}
 
-	const item = queriedState.items[ context ]?.[ key ];
+	const item = queriedState.items[ context ]?.[ recordId ];
 	if ( ! item ) {
 		return false;
 	}
@@ -490,27 +483,27 @@ export function hasEntityRecord(
 /**
  * Returns the Entity's record object by key. Doesn't trigger a resolver nor requests the entity records from the API if the entity record isn't available in the local state.
  *
- * @param state State tree
- * @param kind  Entity kind.
- * @param name  Entity name.
- * @param key   Record's key
+ * @param state    State tree
+ * @param kind     Entity kind.
+ * @param name     Entity name.
+ * @param recordId Record ID.
  *
  * @return Record.
  */
 export function __experimentalGetEntityRecordNoResolver<
 	EntityRecord extends ET.EntityRecord< any >,
->( state: State, kind: string, name: string, key: EntityRecordKey ) {
-	return getEntityRecord< EntityRecord >( state, kind, name, key );
+>( state: State, kind: string, name: string, recordId: EntityRecordKey ) {
+	return getEntityRecord< EntityRecord >( state, kind, name, recordId );
 }
 
 /**
  * Returns the entity's record object by key,
  * with its attributes mapped to their raw values.
  *
- * @param state State tree.
- * @param kind  Entity kind.
- * @param name  Entity name.
- * @param key   Record's key.
+ * @param state    State tree.
+ * @param kind     Entity kind.
+ * @param name     Entity name.
+ * @param recordId Is omitted for keyless entities.
  *
  * @return Object with the entity's raw attributes.
  */
@@ -519,7 +512,7 @@ export const getRawEntityRecord = createSelector(
 		state: State,
 		kind: string,
 		name: string,
-		key: EntityRecordKey
+		recordId?: EntityRecordKey
 	): EntityRecord | undefined => {
 		logEntityDeprecation( kind, name, 'getRawEntityRecord' );
 
@@ -527,7 +520,7 @@ export const getRawEntityRecord = createSelector(
 			state,
 			kind,
 			name,
-			key
+			recordId
 		);
 		const config = getEntityConfig( state, kind, name );
 		if ( ! record || ! config?.rawAttributes?.length ) {
@@ -538,15 +531,15 @@ export const getRawEntityRecord = createSelector(
 		// we return those from record selectors to make rendering,
 		// comparisons, and joins with edits easier.
 		return Object.fromEntries(
-			Object.keys( record ).map( ( _key ) => {
-				if ( config.rawAttributes.includes( _key ) ) {
-					const rawValue = record[ _key ]?.raw;
+			Object.keys( record ).map( ( key ) => {
+				if ( config.rawAttributes.includes( key ) ) {
+					const rawValue = record[ key ]?.raw;
 					return [
-						_key,
-						rawValue !== undefined ? rawValue : record[ _key ],
+						key,
+						rawValue !== undefined ? rawValue : record[ key ],
 					];
 				}
-				return [ _key, record[ _key ] ];
+				return [ key, record[ key ] ];
 			} )
 		) as EntityRecord;
 	},
@@ -554,7 +547,7 @@ export const getRawEntityRecord = createSelector(
 		state: State,
 		kind: string,
 		name: string,
-		recordId: EntityRecordKey,
+		recordId?: EntityRecordKey,
 		query?: GetRecordsHttpQuery
 	) => {
 		const context = query?.context ?? 'default';
@@ -861,7 +854,7 @@ export const __experimentalGetEntitiesBeingSaved = createSelector(
  * @param state    State tree.
  * @param kind     Entity kind.
  * @param name     Entity name.
- * @param recordId Record ID.
+ * @param recordId Is omitted for keyless entities.
  *
  * @return The entity record's edits.
  */
@@ -869,11 +862,11 @@ export function getEntityRecordEdits(
 	state: State,
 	kind: string,
 	name: string,
-	recordId: EntityRecordKey
+	recordId?: EntityRecordKey
 ): Optional< any > {
 	logEntityDeprecation( kind, name, 'getEntityRecordEdits' );
 	return state.entities.records?.[ kind ]?.[ name ]?.edits?.[
-		recordId as string | number
+		String( recordId )
 	];
 }
 
@@ -887,7 +880,7 @@ export function getEntityRecordEdits(
  * @param state    State tree.
  * @param kind     Entity kind.
  * @param name     Entity name.
- * @param recordId Record ID.
+ * @param recordId Is omitted for keyless entities.
  *
  * @return The entity record's non transient edits.
  */
@@ -896,7 +889,7 @@ export const getEntityRecordNonTransientEdits = createSelector(
 		state: State,
 		kind: string,
 		name: string,
-		recordId: EntityRecordKey
+		recordId?: EntityRecordKey
 	): Optional< any > => {
 		logEntityDeprecation( kind, name, 'getEntityRecordNonTransientEdits' );
 		const { transientEdits } = getEntityConfig( state, kind, name ) || {};
@@ -911,9 +904,16 @@ export const getEntityRecordNonTransientEdits = createSelector(
 			return acc;
 		}, {} );
 	},
-	( state: State, kind: string, name: string, recordId: EntityRecordKey ) => [
+	(
+		state: State,
+		kind: string,
+		name: string,
+		recordId?: EntityRecordKey
+	) => [
 		state.entities.config,
-		state.entities.records?.[ kind ]?.[ name ]?.edits?.[ recordId ],
+		state.entities.records?.[ kind ]?.[ name ]?.edits?.[
+			String( recordId )
+		],
 	]
 );
 
@@ -924,7 +924,7 @@ export const getEntityRecordNonTransientEdits = createSelector(
  * @param state    State tree.
  * @param kind     Entity kind.
  * @param name     Entity name.
- * @param recordId Record ID.
+ * @param recordId Is omitted for keyless entities.
  *
  * @return Whether the entity record has edits or not.
  */
@@ -932,7 +932,7 @@ export function hasEditsForEntityRecord(
 	state: State,
 	kind: string,
 	name: string,
-	recordId: EntityRecordKey
+	recordId?: EntityRecordKey
 ): boolean {
 	logEntityDeprecation( kind, name, 'hasEditsForEntityRecord' );
 	return (
@@ -949,7 +949,7 @@ export function hasEditsForEntityRecord(
  * @param state    State tree.
  * @param kind     Entity kind.
  * @param name     Entity name.
- * @param recordId Record ID.
+ * @param recordId Is omitted for keyless entities.
  *
  * @return The entity record, merged with its edits.
  */
@@ -958,7 +958,7 @@ export const getEditedEntityRecord = createSelector(
 		state: State,
 		kind: string,
 		name: string,
-		recordId: EntityRecordKey
+		recordId?: EntityRecordKey
 	): ET.Updatable< EntityRecord > | false => {
 		logEntityDeprecation( kind, name, 'getEditedEntityRecord' );
 		const raw = getRawEntityRecord( state, kind, name, recordId );
@@ -979,7 +979,7 @@ export const getEditedEntityRecord = createSelector(
 		state: State,
 		kind: string,
 		name: string,
-		recordId: EntityRecordKey,
+		recordId?: EntityRecordKey,
 		query?: GetRecordsHttpQuery
 	) => {
 		const context = query?.context ?? 'default';
@@ -990,7 +990,9 @@ export const getEditedEntityRecord = createSelector(
 			]?.[ recordId ],
 			state.entities.records?.[ kind ]?.[ name ]?.queriedData
 				.itemIsComplete[ context ]?.[ recordId ],
-			state.entities.records?.[ kind ]?.[ name ]?.edits?.[ recordId ],
+			state.entities.records?.[ kind ]?.[ name ]?.edits?.[
+				String( recordId )
+			],
 		];
 	}
 );
@@ -1023,7 +1025,7 @@ export function isAutosavingEntityRecord(
  * @param state    State tree.
  * @param kind     Entity kind.
  * @param name     Entity name.
- * @param recordId Record ID.
+ * @param recordId Is omitted for keyless entities.
  *
  * @return Whether the entity record is saving or not.
  */
@@ -1031,12 +1033,12 @@ export function isSavingEntityRecord(
 	state: State,
 	kind: string,
 	name: string,
-	recordId: EntityRecordKey
+	recordId?: EntityRecordKey
 ): boolean {
 	logEntityDeprecation( kind, name, 'isSavingEntityRecord' );
 	return (
 		state.entities.records?.[ kind ]?.[ name ]?.saving?.[
-			recordId as EntityRecordKey
+			String( recordId )
 		]?.pending ?? false
 	);
 }
@@ -1060,7 +1062,7 @@ export function isDeletingEntityRecord(
 	logEntityDeprecation( kind, name, 'isDeletingEntityRecord' );
 	return (
 		state.entities.records?.[ kind ]?.[ name ]?.deleting?.[
-			recordId as EntityRecordKey
+			String( recordId )
 		]?.pending ?? false
 	);
 }
@@ -1071,7 +1073,7 @@ export function isDeletingEntityRecord(
  * @param state    State tree.
  * @param kind     Entity kind.
  * @param name     Entity name.
- * @param recordId Record ID.
+ * @param recordId Is omitted for keyless entities.
  *
  * @return The entity record's save error.
  */
@@ -1079,11 +1081,12 @@ export function getLastEntitySaveError(
 	state: State,
 	kind: string,
 	name: string,
-	recordId: EntityRecordKey
+	recordId?: EntityRecordKey
 ): any {
 	logEntityDeprecation( kind, name, 'getLastEntitySaveError' );
-	return state.entities.records?.[ kind ]?.[ name ]?.saving?.[ recordId ]
-		?.error;
+	return state.entities.records?.[ kind ]?.[ name ]?.saving?.[
+		String( recordId )
+	]?.error;
 }
 
 /**